+++ workspace/solutions/kotlin/lib/src/test/kotlin/s0/s7/s3/s9/s_0739_test.kt
@@ -1,0 +1,16 @@
+package s0.s7.s3.s9
+
+import utils.T
+import kotlin.test.Test
+import s0.s7.s3.s9.s_0739_00.Solution as s_0739_00
+
+class s_0739_test {
+    @Test
+    fun test() {
+        val t = T(
+            this::class,
+            s_0739_00::class,
+        )
+        t.run()
+    }
+}