+++ workspace/solutions/rust/src/s0/s7/s3/s9/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_0739_00;
+pub mod s_0739_test;
+
+pub type TSolution = fn(Vec<i32>) -> Vec<i32>;